requirements.txt
python-dotenv
langchain
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
psycopg2-binary
pinecone-client
pgvector
streamlit
google-generativeai
langchain-google-genai
pdfkit
wkhtmltopdf
streamlit-javascript
google-cloud-aiplatform
langchain-openai
